+++ js/main.js
@@ -1,5 +1,5 @@
 import {receivePostsAsync} from './network.js';
 import {renderThumbnails} from './thumbnails.js';
-import './uploadForm.js';
+import './photouploading/form.js';
 
 receivePostsAsync(renderThumbnails);
