+++ js/photouploading/common.js
@@ -1,0 +1,2 @@
+export const form = document.querySelector('.img-upload__form');
+export const imagePreview = form.querySelector('.img-upload__preview');
